looks as if it wokrs, some erros are still there, i guess is overflow of datatypes, todo: convet double -> long double

+++ include/dataPoints.h
@@ -29,7 +29,7 @@ struct DataPoints
 	int num_features;
 };
 
-DataPoints *AllocateDataPoints(int num_features, int num_data_points, bool malloc_managed = true);
+DataPoints *AllocateDataPoints(int num_features, long num_data_points, bool malloc_managed = true);
 void DeallocateDataPoints(DataPoints* data_points);
 float Distance(DataPoints *p1, DataPoints *p2, int point_id, int cluster_id);
 float MeanSquareError(DataPoints *point, DataPoints *centroid);
